moved around some code
